fixed import error bug a bit (now import * imports everthing including stuff that should maybe be hidden... but I think that's better than breaking the behaviour of __name__@
